refactor: 툴팁 관련 로직 훅으로 분리
