+++ packages/ui/Tooltip/Tooltip.tsx
@@ -1,49 +1,30 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { BubblePointIcon, AlertIcon } from 'Tooltip/icons';
 import * as S from './style.css';
+import useTooltip from 'Tooltip/useTooltip';
+import { HTMLAttributes } from 'react';
 
-interface TooltipProps {
+interface TooltipProps extends HTMLAttributes<HTMLDivElement> {
   triggerContent: string;
-  children?: React.ReactNode;
 }
 
-const Tooltip = ({ triggerContent, children }: TooltipProps) => {
-  const [isVisible, setIsVisible] = useState(false);
-  const tooltipRef = useRef<HTMLDivElement | null>(null);
-
-  const handleMouseEnter = () => {
-    setIsVisible(true);
-  };
-
-  const handleMouseLeave = () => {
-    setIsVisible(false);
-  };
-
-  useEffect(() => {
-    const tooltip = tooltipRef.current;
-
-    if (tooltip) {
-      tooltip.addEventListener('mouseenter', handleMouseEnter);
-      tooltip.addEventListener('mouseleave', handleMouseLeave);
-
-      return () => {
-        tooltip.removeEventListener('mouseenter', handleMouseEnter);
-        tooltip.removeEventListener('mouseleave', handleMouseLeave);
-      };
-    }
-  }, []);
+const Tooltip = ({ triggerContent, children, ...props }: TooltipProps) => {
+  const { isVisible, tooltipRef } = useTooltip();
 
   return (
-    <div role='tooltip' ref={tooltipRef} className={S.tooltipWrapper}>
-      <span className={S.trigger}>
+    <div className={S.tooltipWrapper} ref={tooltipRef} {...props}>
+      <span className={S.trigger} aria-describedby={isVisible ? 'tooltip-content' : undefined} tabIndex={0}>
         {triggerContent} <AlertIcon />
       </span>
-      {isVisible && (
-        <div className={S.contentWrapper}>
-          <BubblePointIcon className={S.bubblePointIcon} />
-          <span className={S.content}>{children}</span>
-        </div>
-      )}
+      <div
+        role='tooltip'
+        className={`${S.contentWrapper} ${S.contentWrapperVariant[isVisible ? 'visible' : 'hidden']}`}
+        data-visible={isVisible}
+        aria-hidden={!isVisible}
+      >
+        <BubblePointIcon className={S.bubblePointIcon} />
+        <span className={S.content}>{children}</span>
+      </div>
     </div>
   );
 };
